requirements.txt
aiogram
ollama
python-dotenv
telethon
chromadb
sqlalchemy
aiosqlite
greenlet
